+++ Release Notes.md
@@ -1,6 +1,7 @@
 ### New in 0.2
 
 * Remove CORS Specific headers
+* Fix bug where connection was dropped when no message was transmitted
 
 ### New in 0.1