+++ Release Notes.md
@@ -2,6 +2,7 @@
 
 * Remove CORS Specific headers
 * Fix bug where connection was dropped when no message was transmitted
+* Correctly handle errors from the engine.io layer in the socket.io layer, closing the socket.
 
 ### New in 0.1